merge-lock-jenkins-service/requirements.txt
requests==2.2.1
python-dotenv==0.6.1
